Refactored for consistency Separated out the semester logic into its own service Renamed campus-service asu-campus-service like all the rest of the asu specific services
